Move tab_transformer model to appropriate dir. Fix imports.
